Reapply "사라진 파일들 다시 받아서 백업"
This reverts commit 0590b46719101ced2b65f145ecbe982387c5c108.

+++ moneytree_back/src/main/java/com/moneytree_back/domain/DepositTermination.java
@@ -5,6 +5,7 @@ import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -31,10 +32,10 @@ public class DepositTermination {
     private TerminationType terminationType;
 
     @Column(name = "termination_rate", nullable = false, precision = 5, scale = 2)
-    private Double terminationRate;
+    private BigDecimal terminationRate;
 
     @Column(name = "termination_final_amount", nullable = false, precision = 15, scale = 2)
-    private Double terminationFinalAmount;
+    private BigDecimal terminationFinalAmount;
 
     @Column(name = "termination_date", nullable = false)
     private LocalDate terminationDate;
